# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manages books, series, and universes for the Lexi platform.",
        "title": "Lexi Book Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/series": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists series, optionally filtered by universe or position within the universe",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "List series",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by universe ID",
                        "name": "universeId",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by exact position in the universe",
                        "name": "universePosition",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by position \u003c= this value",
                        "name": "maxUniversePosition",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/schemas.SerieRead"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Create a serie",
                "parameters": [
                    {
                        "description": "Serie to create",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.SerieCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/schemas.SerieRead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/series/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get a serie by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Serie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.SerieRead"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Update a serie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Serie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schemas.SerieUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.SerieRead"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "series"
                ],
                "summary": "Delete a serie",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Serie ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/schemas.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schemas.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "schemas.SerieCreate": {
            "type": "object",
            "required": [
                "serie"
            ],
            "properties": {
                "serie": {
                    "type": "string"
                },
                "universeId": {
                    "type": "string"
                },
                "universePosition": {
                    "type": "integer"
                }
            }
        },
        "schemas.SerieRead": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "serie": {
                    "type": "string"
                },
                "universeId": {
                    "type": "string"
                },
                "universePosition": {
                    "type": "integer"
                }
            }
        },
        "schemas.SerieUpdate": {
            "type": "object",
            "properties": {
                "serie": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Raw JWT issued by lexi-users-ms's /login (no \"Bearer \" prefix).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}